[Fixes #13906693] Make GetPendingReserveOnItem more selective
The GetPendingReserveOnItem function created for #8315135 was
being too inclusive in its query, triggering the appearance
of item-level holds in the detail view even when there were
none. This patch changes it so it only returns a defined value
if there is an item-level hold or the item has an associated
reserve that is Waiting or in Transit.