PT13051737: SCLS and Pioneer debt collect ignoring --max-wait parameter
The result of this bug was solved by adding an -ignore command line
parameter switch to the amountdue_notices.pl script.  This takes a
number of days argument and excludes from the amount owed summation
any transaction older than the date provided.  This exclusion results
in the borrowers.amount_nofity_date potentially not being set.