[Fixes #12849447] Clarify when item unavailable for SCO renewal
UI ambiguity was contributing to a situation where it was easy for
a patron to think they were renewing an item when in fact they
were choosing to indicate they had returned it. A working button
in a table column labeled "Renew" should only be able to renew
items, rather than quietly presenting the very similar option to
"Return Item". This patch disables one-click returns and makes
the item's unavailability more visually apparent.