PT13497871:  Blank lines in TalkingTech MESSAGE file
This patch cleans up some of the original coding for creating the
Talking Tech MESSAGE file.  Originally, the code merely appended a
new line onto a file contained in /tmp.  However, as a result, of
file clobbering and the setup of the operational environment, the
process was changed to create single line files into a new more
stable directory referenced by TalkingTechMessagePath.

Some additional cleanup of advance_notices.pl and overdue_notices.pl
was done as well.