[Fixes #14125571] Optimize checkout by reducing patron detail
The amount of information the patron info box at the bottom
of the circ window compiles is too great for a speedy
checkout process if the patron has many items on reserve,
especially if those reserves are on bibs with lots of items
and/or bibs that have lots of reserves.

This eliminates the patron_infobox and streamlines acquisition
of many of the other details about the patron account. It also
replaces the infobox with a listing of the item that was just
checked out after a successful issue.