Add Koha.pm and include it universally
At this point all we're doing is turning on warnings, strict,
and 5.10 features by default in all files. This also defines
a $Koha::VERSION variable which is going to be the basis for
the code revision number, as opposed to $C4::Context::VERSION,
which is the data config version number.