Fixes to SIP(2) checkin response message (10)
The SIP checkin response message was producing some erroneous alerts in
a couple of cases.  Specifically, an alert was being produced on a suspended
hold as well as an in-transit hold that was being transferred to another
branch.  This caused some problems with ITG's sorting equipment.

This patch also changes the extended SIP2 DA field to only include the
patron name rather than the name, email, and phone number is was currently
set up to provide.