De-globalize $dbh in C4::VirtualShelves
The package global assignment of $C4::VirtualShelves::dbh, in
addition to being just plain shoddy, is not persistence-safe. It
also demands a database connection at module load time, which is
a dependency we need to avoid both for testing and for the Plack
localization middleware.

This also removes a couple of other stray $dbh initializations
and other cruft in other files.