Allow for dynamic multi-instance servicing
By handing off a hash of hostname-to-KOHA_CONF mappings to the
Koha::Plack::Localize middleware via the host_map parameter,
a single running Plack instance can be used to service multiple
installations for distinct libraries. The absence of this
parameter causes the traditional $ENV{KOHA_CONF} to be used.

It may be useful in the future to also add a "host_mapper"
parameter which could be handed a coderef that would return
the appropriate config file based on any number of factors.