PT13633157: OPAC not obeying 942$n bib suppression setting (SC #14321)
If the framework allows the biblionumber (999$c) field to be displayed and
edited within the MARC editor, then this results in an additional 999$c value
being added to the MARC record.  This is because the code already pulled the
biblionumber from the biblio table.  This patch removes the addition of the
extra 999$c field, although this field shouldn't really be editable in the
first place.