[delivers #16029551] COinS calculator dying on unchecked return value
If, as in the case of a newly deleted title, an OPAC search result list
includes a title that no longer exists, the COinS creator would die
from accessing an unchecked return value. This patch explicitly croaks
in that case and places catch()es within the caller to emit a simple
warn instead of dying.