PT16310341: SIP2 differentiation between item-level and title-level holds
The SIP2 code was essentially forcing a priority hold regardless of whether
it was a title-level or item-level hold.  Altered the code to account for
an item level hold by testing for an itemnumber and non-found status in the
reserves table.  Needed to regain patron info if an item level hold was in
place for an item that was being checked in.