[delivers #16913917] Global namespace pollution in C4::Letters
%C4::Letters::handles can end up bleeding information from one
session to the next when running in a persistent environment.
Caching of this data does not create a material performance
improvement, so this patch just removes it altogether. If it
turns out there are some use cases where caching is indicated,
a new mechansim using CHI handles stored in C4::Context is
required.