PT13498505: Implement TalkingTech's customized 43/44 message pair.
TalkingTech created their own customized SIP2 message/response pair.
This was to provide real time updates of the success or failure of
the outgoing phone calls they were dialing.

This patch implements the acceptance of a 43 message and the creation
of the 44 response.  Ultimately, if a call was not successful, then
a print notice is queued in the message_queue table.