PT15650379: Cannot Save Changes to Leader
When staff edits the leader via the pop-up window, it leaves the
calculated fields blank (including the first five characters).  This
was complicated by the fact that the record status (position 5) was also
blank if a status of 'a' was chosen.  The result was a shortened leader
being processed and subsequently not saved because it was not 24 characters
in length.

This patch applies the record status value of 'a' if selected and adds
padding for the record length field.  A check is also done to make sure
any edits are not coming from the initial editing screen.  If this is the
case, then the padding is not necessary.