[#18149087] (cleanup) Calling exit() from core modules
Not that these two exit() calls were reachable, anyway, since an
unconditional return is called just prior to each of them, but
exit() should never be called from a module. It's safe to call
it from a .pl script, but calling it from within a .pm causes
the Plack engine itself to exit, which is bad news.