[delivers #18667851] Force C4::Context->preference() to return scalar
If a pref is not registered, preference() returns a context-dependent
value.  Lots of code, in particular in C4::Auth::get_user_and_template,
calls preference() from a list context, and that can cause things like
setting values in a hash to have empty lists as values, which ends up
being a non-value, and completely rearranging the var/val relationships
that ought to be set.