[#21283309] Remove display of reading history when disabled
If KeepPreviousBorrower is enabled it can result in an item
remaining displayed on the OPAC reading history page even
if the patron has reading history disabled. It will remain
there until either KeepPreviousBorrowerInterval passes or
the item is checked out by another patron. This patch just
short-circuits the reading history page so that it's not
retrieved at all if the patron has history disabled.