[#21093921] Still allowing on-shelf holds if also available at another branch
There was a discrepancy in the available-for-holds logic that
allowed on-shelf holds from the OPAC if there were copies
available elsewhere, too. This is because it would still be
possible to place an item-specific reserve in the case of
serials, where it could be that you want another issue
altogether. However, this commit fixes this while still
differentiating for serials. This means it's still possible
to do an on-shelf reserve of serial items. There's no way
around that.

The only functional piece in this commit is the call to the
new BranchHasACopy() function. All else is cosmetic to make
the conditional more readable.