Utilize DBIx::Connector for C4::Context
This is to clean up a big chunk of the DBI
connection code in C4::Context by consolidating
(and stabilizing) the management into
DBIx::Connector. It brings the replica_dbh
handling into the fold and automatically uses
the 'fixup' mode to obviate the expensive
'database_paranoia_mode' previously in place.