[#14843887] staff advanced search stripping letter x from ISBN search
Modifies earlier code in staff advsearch.tmpl

    q = q.replace(/[^0-9]/g,'');

that was done in PT 6321425 search ISBN dashes. The above replacement was
unduly restrictive and dropped all non-numeric characters; this made it
impossible to search for an ISBN containing an X and narrow in on that
specific record.  Fix leaves alpha and numeric characters alone but strips
hyphens, whitespace, and other characters.