[#18084285] Removes New Patron button from Circ Checkout
The button for adding a new patron under Circ, Checkout was using
patron category_type instead of categorycode to create a new patron
which meant that new patron record defaulted to the first available
Child patron category.  Removing the button from that screen -- it's
not needed and trying to fix the underlying code is difficult.