[#20749431] fix 008 default for Koha authority records
The tag 008 of the stubby authority record was very limited and needed to be
manually linked to a biblio record after the initial creation. This meant if
a another biblio was created that could have used a previous created stubby,
(e.g., same author) it wouldn't -- it created a new stubby, This creates duplicate
authorities. In addition, running link_bibs_to_authorities.pl unlinked all
stubby records because the 008 was not set to link to anything.  This fix sets
proper default values for the 008 so that multiple records can be found and
linked to the same authority record.