[#20674153] Get reading history item type from item
On the staff patron reading history tab, the item type was being
pulled from the MARC record 942c entry rather than the item record's
item type.  This was frequently inaccurate or non-existent, and was
confusing.  The fix changes the source for the display field to
the item record's item type.