[#16567199] Remove Accession Date link from Item tab
In the items tab of a record display, the Accession Date was linking
to the acquisitions order record.  If the site had not used the
acquisitions module, the link was invalid and produced a
DBD::mysql::st execute failed error message.  This patch removes the
link from the items tab.