[#23030301, #23030217] (fixup) Cannot reserve reserved on-shelf items
If there was a copy of a title still on the shelf (having not
yet been picked for a pending reserve), a patron in the branch
could place a reserve on the title. This work turns that around
such that the patron is blocked with a message that the item is
available on the shelf.