[#25161833] (more) Clean up IP address parsing
There are some proxies which insert a non-IP value into
the HTTP_X_FORWARDED_FOR header (e.g. "unknown") to
protect privacy or who knows why. This would bunge up
Koha's IP detection. This patch looks for something that
looks like a good address, then if all else fails it
resorts to reporting that the client is not in any branch.