[#24011971] Enable OPAC search constraints based on category name
Instead of placing the CCL string into the <default_search_limit>
contents, a branchcategory name can be specified in the "group"
attirbute of the tag, as such:

<opachosts>
  <opac hostname="my-opac.kohalibrary.com">
    <default_search_limit group="GROUP1" override="1" />
  </opac>
</opachosts>

This will prompt Koha to set default search limits, selecting
only titles having items available at branches in the given
branch category.

Search results and title details can be further limited to
display only information about items that are part of the
search domain of the current agent. This is accomplished by
adding a 'hide_other_items="1"' attribute to the tag, as:

    <default_search_limit group="GROUP1"
      hide_other_items="1"
      override="1" />