[#20608005/#25662245] Item-level holds not getting fulfilled
Yet another instance of what amounts to a "type" error
emerging from the arbitrary and inconsistent naming
of data members. In this case it was that a data member
named "borrowercategory" was being passed to GetIssuingRule(),
but the data passed into the calling function had that value
stored as its column name, "categorycode".

Haskell is looking better all the time.