[#25974505] Disable patron group checking in reserves handling
In situations where a bib had reserves on it by patrons
belonging to branches not in the current agent's branch group,
CheckReserves would trigger an exception by eventually calling
GetCircControlBranch with an undefined borrower category.

This is not a nice solution, but it will do until we can
rethink the permissions architecture for how patron hiding
can work.