[Fixes #24595459] patron record creation- auto calculate not working as expected
Added a date-calcuation check to memberentry.pl for ops insert,modify,save.
Also added a preparatory conversion of the required base date, dateenrolled, to iso
format before calling GetExpiryDate() to calculate the expiry date. This is the date
that the story reported was not correctly calculated.

Removed two prior unnecessary calls to GetExpiryDate in the code that also had
no pre-check and conversion to iso of the dateenrolled date.