[#24985605] Add identifiers to various sections of OPAC
Sponsored by National Association of Realtors.  Adds unique
identifiers to various places in the OPAC so they can be
changed or hidden with a jquery and controlled by separate
stylesheet entries.  Identifiers include separate IDs for each
left-side tab on the patron OPAC account, Holds tab and table
headers, Pickup location table headers, Place Hold wording in
Cart, and empty placeholder IDs which can be used for custom
messages on the account Checked Out, Overdue, and Holds tab
(also on the Confirm Holds screen).

In addition, created unique identifiers for the labels and contents
of many fields in the XSLT results list and title display screens.
This allows sites to reword the labels, change the font/color, or
hide a particular label and field altogether.

See separate documentation for details on usage and syntax.