[#10192207] Fix for fines script stalling on deleted items
If an item was checked out and deleted without being checked in,
a null itemnumber in the issues table made the fines script stop
processing without any alerts.  Any entries in the issues table
after the orphaned record did not accrue fines.  This adds a fix
to skip entries with null itemnumbers.