[#27529753] Retain limit setting when sorting Reading History
The OPAC Reading History tab has options to display all or
last 50 entries, and to sort by Title (if sorted, sort back
by date).  However, sorting by Title/Date ignored the display
all selection, and only used the last 50 entries.  This makes
the sorts obey the limit parameter.

Note that if the limiter is last 50, the list returned will
be different depending on whether title or date is sorted. If
by title, the selection is alpha; if by date the selection is
most recent 50.