[#30164727] [#30164771] Expired/canceled holds availability.
When a hold is canceled or expires while on the holds shelf (i.e.
`found` eq 'W'), it will now appear as 'On hold' in the OPAC iff there
are further holds placed on the title or item.  Previously these would
appear as available.  This edge case is important when there are many
items and many holds.  As a side effect, we backfill `found` values,
adding statuses for Canceled and Expired to more readily identify state
in old_reserves.