[#32857887] Clarify which maxfine setting to apply
If both the circ rule and the MaxFine syspref had values for
maximum fine, the system was applying the lesser. This meant
that in some cases the circ rule value was not being used (for
example, a circ rule max fine of $10 but a syspref of $5).
Modifies the code so that if the circ rule has a value for
max fine, that value is used regardless of the syspref contents.