[#38767751] Add batch upload resiliency to solrsync
The solrsync tool uploads renders documents in batches (of 100
by default). If one of those turns out to be rejected by the Solr
server for a schema violation or some such, the entire batch gets
rejected. This patch modifies the failue handling so that
solrsync splits the batch and recursively retries the upload with
smaller and smaller portions of the batch, failing only on the
documents that are specifically objected to by the Solr schema.

Secondarily, it renames the "--batch_size" command line parameter
to "--batch-size" to conform with convention.