[#40662077] SIP2: enable "issue" message to perform a renewal
3M SIP machines don't send a "renew" packet for items already
checked out, they send an "issue" packet. Koha will reject
these requests since the item is already checked out. This
is because of the potential for some types of machines to do
multiple scans of items, thus creating unwanted renewals.

This commit adds a SIP_RenewOnIssue syspref which when
enabled permits this behavior.