[#30168727] Display session checkouts on circ screen
The list of "session checkouts" is derived from looking at the
list of all checkouts made to the same patron within the past
10 minutes. This allows the librarian to briefly interrupt the
checkout flow (for a search, question about fines, whatever)
and return to the circ screen still seeing the current issue
stack.

However if a patron, for example, were to checkout out a stack
of books with a self-check station and then go to the counter
to check out a special resrve book, the librarian's screen will
still show the entire stack of books even though that librarian
did not personally check them out.