Fix edge case that could crash solrsync
Testing the conditional "unless $doc" causes $doc
to be tested as a string value, which in turn
causes the stringify operator to unroll the doc data
into and XML string. If there's an error in rendering
the XML (char encoding, etc.), XML::Easy will die.

Even in the successful case, turning the doc to XML
is expensive. Testing for definedness rather than truth
is cheap and won't die.