[#39137907] Include extended attributes in patron searches
Using the 'SearchExtendedPatronAttributes' syspref, libraries can
opt to include extended attribute values when conducting patron
searches via C4::SearchMember(). This includes the main page
search, nav bar search, circ search, and possibly other search
interfaces.

Note that this option potentially has a very high cost associated
with it in that the borrower_attributes.attributes values are not
indexed. A full table search is required for these searches. The
time required to complete a search query will increase linearly
with the number of staff_searchable rows in the table.