[#45854197] Workaround for EBSCO's broken EDS availability checker
Activation of this feature is done by creating a config file option:

<option name="explicit_availability">STRING</option>

The value of "STRING" can be whatever value or even a zero-length string.
The value will be set as the value of the $q subfield of any 952 tags
that have *no* $q already. If the tag has an existing $q subfield, this
option will make no change.

c.f. LAK #41832679