[#51023687] Add SearchByRcn syspref
Turning on this sypref (the default) will cause Koha to link
heading searches of authors and subjects with the linked_rcn
field. Disabling the pref employs a value search.

Note that this does not address the existing bug where value
searches are constructed using only the $a rather than the
whole field.