[#50884419] Lots of fixes to syspref handling and normalization
Modifies systempreferences DB schema to store only
variable/value pairs. All other fields are derived from
the values in the defaults file.

Adds a "tags" value to each variable. This is a list
of tabs on which a pref should be included in the prefs
editor. Currently only the first value is supported.

Prefs are no longer case insensitive. The update script
makes an effort to normalize syspref names to the
canonical version.

Removes abililty for the editor to add new prefs, delete
prefs, or edit their type, explanation, options, etc.
There was never a legit use case for this functionality,
and it has the potential to cause considerable trouble.

Attempts to retrieve bogus prefs now result in croaks
rather than warnings.

It's now possible to change a syspref's description
without running incrementing the db schema version.