[#44086709] Reading history retention - preference in patron record not being followed
The main issue was the anonymise_previous_borrowers.pl cron script not
being run when the KeepPreviousBorrower syspref was turned ON.

There was, however, an issue with the placement of the call to the
AnonymiseIssueHistory subroutine.  It was occurring before the latest
return was _MarkIssueReturned.  As a result, it was still in the issues
table and not immediately anonymized.