[#55027870] [#53724683] Correct summing error in holds-per-itemtype limits.
This commit corrects one error, but leaves a second unfixed.
In the OPAC, certain conditions will cause error messages to
be suppressed, and a hold to not be placed without displaying
an error.  This happens when the AllowMultipleHoldsPerBib
system preference is in effect.  Level of effort for fixing
that bug is substantial, so hopefully its impact will be fairly
small now that the summing error is fixed.