[#42084895] Some overdue notices not instantiating
Some overdue notices were not being emailed or printed when the
TalkingTechEnabled syspref was turned ON, but the letter.ttcode
was not set. Essentially this was a problem with a consortium
that only had certain branches uses the Talking Tech software.

This patch also corrects the update or insert of the ttcode value
into the database to use the correct absence-of-value NULL rather
than a string "NULL".