[#55643024] Force scalar context in return of language list
Somehow CHI::compute() has changed semantics such that in a
list context it returns only the first element of the language
list rather than the whole list. This is strange principally
because the language list is a ref and should therefore always
be a scalar. CHI appears to be opting to dereference the
arrayref when called in a list context.