[#55919622] Enhance SIP2 to prevent the use of SCs by expiring patrons
This patch will deny a patron's ability to use a self check machine when
their account is going to expire in a specific number of days.  This number
is configurable via the SIPExpiryPeriod system preference.  In addition, the
message that is sent to the SC screen can be configured via the
SIP_EXPIRY_ALERT notice.