[#57969724] Merge authority records
Authorities popping up in search results can be stored
in a basket. Examining the contents of the basket allows
the user to merge all the authorities into a single one.
What this boils down to is that the bibs linked by each
of the unkept bibs have their fields rewritten such that
any field linked to the un-kept authority is now linked
to the kept authority.