[#58706448] Remove yui widget in patron category select.
For library systems with large numbers of patron categories,
the yui widget for selecting patron category to add a new patron
was slow and obstructed workflow.  This patch replaces it with
a standard pulldown select box.

This change also deprecates the system preference AddPatronLists,
which did not work as advertised.