[#53890731] Fully assert sort order in patron search results
MySQL conveniently has some variation in its retrieval of
WHERE..IN clause results. Some versions will preserve the
order of the elements given in the IN clause. Others do not
and will leave the results unsorted. Because of the two-step
SELECT process introduced by the work in #39137907, this
would cause screenfuls of results to be jumbled in the
latter case, though all the entries in screen one would be
earlier than all the entries in screen two, for example.

Anyway, imposing the same sort on the subsequent result set
fixes the issue.