[#60764316] Enhance SIP2 for integration with Envisionware
This patch utilizes the SIP2 protocol to provide the ability for patrons to
make payments on their outstanding balances.  Specifically it was structured
to work with Envisionware's e-commerce software solution.

The primary need is to configure the EWSipUser system preference with a
specific Koha SIP user.  This account must also be entered into the
SIPconfig.xml file.  The reason for this is to provide the necessary
output specifications for the EnvisionWare software.