[#66465058] Add additional Talking Tech notice types
Only a subset of the available Talking Tech notice types were able to
be selected from the drop-down menu when adding or modifying a notice.
This patch adds the remaining TT notice type options.

Additionally, the notice types were originally hard-coded into the template.
This patch refactors this code and utilizes a syspref to define the
available notice type options.