[#70787702] [#70787786] Alternative calculation of amount owed to compare to blocking and notice thresholds
A new syspref, ExcludeAccruingInTotal, was created to allow for an
alternative calculation of the amount a patron owes when compared to
the holds block, circ block, and billing notice thresholds.  Previously,
this amount included outstanding as well as accruing fees minus any
credits.  If the ExcludeAccruingInTotal syspref is turned ON, then this
calculation removes the accruing fees from the total due calculation.