[#91767014] Pay all button will allow the total outstanding to be paid.
The first commit has only style changes. The second commit has the bug
fix. In certain situations a fee will have been paid but appear as not
paid. In order to remedy this it is necessary to reverse the charge that
paid the fee and then use the pay all functionality. This bug fix will
stop the pay all button not working at all in such situations. Now it
calculates the amount to pay and matches the total outstanding.