[#87746512] Disallow in transit items as available for holds
When an item is being transferred, it is counted as available for
filling holds. This patch disallows that behavior by including a check
to make sure an item for a hold is not in the state of being transferred
from one branch to another. This ensures that only 'Available' status
items are used to fulfill holds.